split out common codec elements and renamed Concat to Flatten
