+++ encoding/cbf/common_codec/bool.go
@@ -1,0 +1,61 @@
+/*
+ * Cadence - The resource-oriented smart contract programming language
+ *
+ * Copyright 2019-2022 Dapper Labs, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package common_codec
+
+import (
+	"fmt"
+	"io"
+)
+
+type EncodedBool byte
+
+const (
+	EncodedBoolUnknown EncodedBool = iota
+	EncodedBoolFalse
+	EncodedBoolTrue
+)
+
+func EncodeBool(w io.Writer, boolean bool) (err error) {
+	if boolean {
+		_, err = w.Write([]byte{byte(EncodedBoolTrue)})
+	} else {
+		_, err = w.Write([]byte{byte(EncodedBoolFalse)})
+
+	}
+
+	return
+}
+
+func DecodeBool(r io.Reader) (boolean bool, err error) {
+	b := make([]byte, 1)
+	_, err = r.Read(b)
+	if err != nil {
+		return
+	}
+
+	if EncodedBool(b[0]) == EncodedBoolTrue {
+		boolean = true
+	} else if EncodedBool(b[0]) == EncodedBoolFalse {
+		boolean = false
+	} else {
+		err = CodecError(fmt.Sprintf("invalid boolean value: %d", b[0]))
+	}
+
+	return
+}
