+++ encoding/cbf/common_codec/location.go
@@ -19,262 +19,11 @@
 package common_codec
 
 import (
-	"encoding/binary"
 	"fmt"
 	"io"
-	"math/big"
 
-	"github.com/onflow/cadence"
 	"github.com/onflow/cadence/runtime/common"
 )
-
-//
-// LengthyWriter
-//
-
-type LengthyWriter struct {
-	w      io.Writer
-	length int
-}
-
-func NewLengthyWriter(w io.Writer) LengthyWriter {
-	return LengthyWriter{w: w}
-}
-
-func (l *LengthyWriter) Write(p []byte) (n int, err error) {
-	n, err = l.w.Write(p)
-	l.length += n
-	return
-}
-
-func (l *LengthyWriter) Len() int {
-	return l.length
-}
-
-//
-// LocatedReader
-//
-
-type LocatedReader struct {
-	r        io.Reader
-	location int
-}
-
-func NewLocatedReader(r io.Reader) LocatedReader {
-	return LocatedReader{r: r}
-}
-
-func (l *LocatedReader) Read(p []byte) (n int, err error) {
-	n, err = l.r.Read(p)
-	l.location += n
-	return
-}
-
-func (l *LocatedReader) Location() int {
-	return l.location
-}
-
-//
-// Bool
-//
-
-type EncodedBool byte
-
-const (
-	EncodedBoolUnknown EncodedBool = iota
-	EncodedBoolFalse
-	EncodedBoolTrue
-)
-
-func EncodeBool(w io.Writer, boolean bool) (err error) {
-	b := EncodedBoolFalse
-	if boolean {
-		b = EncodedBoolTrue
-	}
-
-	_, err = w.Write([]byte{byte(b)})
-	return
-}
-
-func DecodeBool(r io.Reader) (boolean bool, err error) {
-	b := make([]byte, 1)
-	_, err = r.Read(b)
-	if err != nil {
-		return
-	}
-
-	switch EncodedBool(b[0]) {
-	case EncodedBoolFalse:
-		boolean = false
-	case EncodedBoolTrue:
-		boolean = true
-	default:
-		err = CodecError(fmt.Sprintf("invalid boolean value: %d", b[0]))
-	}
-
-	return
-}
-
-//
-// Length
-//
-
-// TODO encode length with variable-sized encoding?
-//      e.g. first byte starting with `0` is the last byte in the length
-//      will usually save 3 bytes. the question is if it saves or costs encode and/or decode time
-
-// EncodeLength encodes a non-negative length as a uint32.
-// It uses 4 bytes.
-func EncodeLength(w io.Writer, length int) (err error) {
-	if length < 0 { // TODO is this safety check useful?
-		return CodecError(fmt.Sprintf("cannot encode length below zero: %d", length))
-	}
-
-	l := uint32(length)
-
-	return binary.Write(w, binary.BigEndian, l)
-}
-
-func DecodeLength(r io.Reader) (length int, err error) {
-	b := make([]byte, 4)
-
-	bytesRead, err := r.Read(b)
-	if err != nil {
-		return
-	}
-	if bytesRead != 4 {
-		err = CodecError("EOF when reading length")
-		return
-	}
-
-	asUint32 := binary.BigEndian.Uint32(b)
-	length = int(asUint32)
-	return
-}
-
-//
-// Bytes
-//
-
-func EncodeBytes(w io.Writer, bytes []byte) (err error) {
-	err = EncodeLength(w, len(bytes))
-	if err != nil {
-		return
-	}
-	_, err = w.Write(bytes)
-	return
-}
-
-func DecodeBytes(r io.Reader) (bytes []byte, err error) {
-	length, err := DecodeLength(r)
-	if err != nil {
-		return
-	}
-
-	bytes = make([]byte, length)
-
-	bytesRead, err := r.Read(bytes)
-	if err != nil {
-		return
-	}
-	if bytesRead != length {
-		err = CodecError("EOF when reading bytes")
-		return
-	}
-
-	return
-}
-
-//
-// String
-//
-
-func EncodeString(w io.Writer, s string) (err error) {
-	return EncodeBytes(w, []byte(s))
-}
-
-func DecodeString(r io.Reader) (s string, err error) {
-	b, err := DecodeBytes(r)
-	if err != nil {
-		return
-	}
-	s = string(b)
-	return
-}
-
-//
-// Address
-//
-
-func EncodeAddress[Address common.Address | cadence.Address](w io.Writer, a Address) (err error) {
-	_, err = w.Write(a[:])
-	return
-}
-
-func DecodeAddress(r io.Reader) (a common.Address, err error) {
-	bytes := make([]byte, common.AddressLength)
-
-	bytesRead, err := r.Read(bytes)
-	if err != nil {
-		return
-	}
-	if bytesRead != common.AddressLength {
-		err = CodecError("EOF when reading address")
-		return
-	}
-
-	return common.BytesToAddress(bytes)
-}
-
-//
-// U?Int(8,16,32,64)
-//
-
-// TODO use a more efficient encoder than `binary` (they say to in their top source comment)
-
-func EncodeNumber[T int8 | int16 | int32 | int64 | uint8 | uint16 | uint32 | uint64](w io.Writer, i T) (err error) {
-	return binary.Write(w, binary.BigEndian, i)
-}
-
-func DecodeNumber[T int8 | int16 | int32 | int64 | uint8 | uint16 | uint32 | uint64](r io.Reader) (i T, err error) {
-	err = binary.Read(r, binary.BigEndian, &i)
-	return
-}
-
-//
-// BigInt
-//
-
-func EncodeBigInt(w io.Writer, i *big.Int) (err error) {
-	isNegative := i.Sign() == -1
-	err = EncodeBool(w, isNegative)
-	if err != nil {
-		return
-	}
-
-	return EncodeBytes(w, i.Bytes())
-}
-
-func DecodeBigInt(r io.Reader) (i *big.Int, err error) {
-	isNegative, err := DecodeBool(r)
-	if err != nil {
-		return
-	}
-
-	bytes, err := DecodeBytes(r)
-	if err != nil {
-		return
-	}
-
-	i = big.NewInt(0)
-	i.SetBytes(bytes)
-	if isNegative {
-		i.Neg(i)
-	}
-	return
-}
-
-// Location
 
 func EncodeLocation(w io.Writer, location common.Location) (err error) {
 	switch concreteType := location.(type) {
@@ -458,48 +207,4 @@ func DecodeTransactionLocation(r io.Reader, memoryGauge common.MemoryGauge) (loc
 
 	location = common.NewTransactionLocation(memoryGauge, byteArray)
 	return
-}
-
-//
-// Misc
-//
-
-func Concat[T any](deep ...[]T) []T {
-	length := 0
-	for _, b := range deep {
-		length += len(b)
-	}
-
-	flat := make([]T, 0, length)
-	for _, b := range deep {
-		flat = append(flat, b...)
-	}
-
-	return flat
-}
-
-//
-// Testing
-//
-
-type MockWriter struct {
-	ByteToErrorOn int
-	ErrorToReturn error
-	CurrentByte   int
-}
-
-var _ io.Writer = &MockWriter{}
-
-func (m *MockWriter) Write(p []byte) (n int, err error) {
-	currentByte := m.CurrentByte
-	m.CurrentByte += len(p)
-
-	if m.ByteToErrorOn < 0 || // erroring disabled
-		m.ErrorToReturn == nil || // no erroring
-		currentByte > m.ByteToErrorOn || // already errored
-		m.CurrentByte <= m.ByteToErrorOn { // not yet erroring
-		return len(p), nil
-	}
-
-	return 0, m.ErrorToReturn
 }
